+++ lib/marskal_search/getter_methods.rb
@@ -1,50 +1,67 @@
 class MarskalSearch
 
+  def column_expressions
+    @select_depot[:column_details].collect {|c| c[:expression] }
+  end
+
+  def column_names
+    @select_depot[:column_details].collect {|c| c[:name] }
+  end
+
+  def column_headings
+    @select_depot[:column_details].collect {|c| c[:heading]}
+  end
+
 private
+  #this will build the instance Hash variable called @sql_depot
   def column_details(p_select_clause)
     @select_depot[:count_distinct] = ''
     @select_depot[:ready_for_sql] = ''
     @select_depot[:column_details] = []
-    @select_columns = []
     if p_select_clause.is_a?(Array)
       l_smart_columns = p_select_clause.map{|c| c.split_select_column_alias}
     else
       l_smart_columns = p_select_clause.smart_comma_parse_to_array.map{|c| c.split_select_column_alias}
     end
     
-    l_smart_columns.each do |l_attr_array|
+    l_smart_columns.each_with_index do |l_attr_array, idx|
       l_attr = l_attr_array[0].remove_begin_end_char(COLUMN_WRAPPER_CHAR)
-      l_alias = l_attr_array[1].to_s.blank? ? nil : l_attr_array[1].remove_begin_end_char(COLUMN_WRAPPER_CHAR)
+      l_alias = l_attr_array[1].to_s.blank? ? nil :  Utils.format_alias_name(l_attr_array[1]) #l_attr_array[1].remove_begin_end_char(COLUMN_WRAPPER_CHAR)
       l_col = @model.columns_hash[l_attr] rescue nil
       if l_col.nil?
         l_col = @model.columns_hash[l_alias] rescue nil
       end
 
-      @select_depot[:ready_for_sql] += ', ' unless @select_depot[:ready_for_sql].blank?
-      @select_depot[:ready_for_sql] += "#{l_attr}"
-      @select_depot[:ready_for_sql] += " AS #{l_alias}" unless l_alias.to_s.blank?
-
       @select_depot[:count_distinct] += ', ' unless @select_depot[:count_distinct].blank?
       @select_depot[:count_distinct] += "IFNULL(#{l_attr}, '')"
 
       l_current_col = {
-          name:       l_attr,
+          expression: l_attr,
+          name:       l_alias||l_attr,
           ruby_type:  (l_col.nil?) ? :string : l_col.type,   #to_do, only include for MAX info
           sql_type:   (l_col.nil?) ? 'varchar' : l_col.type,
           type:       (l_col.nil?) ? :string : l_col.type,
       }
-      l_current_col[:alias] = l_alias unless l_alias.blank?
       if l_col.nil?
-        l_current_col[:heading]    = l_alias.to_s.blank? ? l_attr : l_alias
+        l_alias ||= "Column #{idx}"
+        l_current_col[:heading]    = l_alias
       else
         l_current_col[:length]     = l_col.limit||(l_col.precision.to_i + l_col.scale.to_i + 1)
         l_current_col[:precision]  = l_col.precision  unless l_col.precision.nil?
         l_current_col[:scale]      = l_col.scale unless l_col.scale.nil?
-        l_current_col[:heading]    = l_col.name.split('_').map{|w|w.humanize}.join(' ') #TODO look for overrides from api call
+        l_current_col[:heading]    = (l_alias||l_col.name).split('_').map{|w|w.humanize}.join(' ') #TODO look for overrides from api call
+
+        #patch for to_json and as_json
+        if l_col.name.downcase == 'type'       #to_json and as_json chops off a db field if it is named 'type' for some reason
+          l_current_col[:name] =  "#{@table_name.singularize.downcase}_type"   #patch for to_json and the type field, we give it an alias with a space after type and that seem sto take care of problem
+        end
 
         if l_col.type == :datetime   #TODO:check for time fields as well
           l_current_col[:length] = 25
-          # l_current_col.except!(:precision, :scale)
+          unless @output_settings[:datetime_format].blank?            # see if we need to reformat the output of the datetime field
+            l_current_col[:expression]  = "DATE_FORMAT(#{l_current_col[:expression]}, '#{@output_settings[:datetime_format]}')"
+            l_current_col[:name] +=  '_formatted'
+          end
         elsif l_col.type == :date
           l_current_col[:length] = 10
         else
@@ -57,8 +74,13 @@ private
           end
         end
       end
+      l_current_col[:name] = Utils.format_alias_name(l_alias||l_current_col[:name])
       @select_depot[:column_details] << l_current_col
-      @select_columns << (l_current_col[:alias]||l_current_col[:name])
+
+      @select_depot[:ready_for_sql] += ', ' unless @select_depot[:ready_for_sql].blank?
+      @select_depot[:ready_for_sql] += "#{l_current_col[:expression]}"
+      @select_depot[:ready_for_sql] += " AS #{ l_current_col[:name]}" unless  l_current_col[:name] == l_current_col[:expression]
+
     end #end mart_columns
     return self
   end